iot-data-adapter/requirements.txt
fastapi
uvicorn
pydantic
confluent-kafka
paho-mqtt
